pyproject.toml
[project]
name = "django-ecommerce-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.5.3",
    "django>=5.2.8",
    "django-celery-beat>=2.8.1",
    "django-ratelimit>=4.1.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djoser>=2.3.3",
    "flower>=2.0.1",
    "python-json-logger>=4.0.0",
]
